% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_ercc_medians}
\alias{calculate_ercc_medians}
\title{Calculate the median expression for each ERCC group}
\usage{
calculate_ercc_medians(tt)
}
\arguments{
\item{tt, }{data.frame, typically the output of the \code{\link[limma:topTable]{limma::topTable()}}
function}
}
\value{
A tibble with one row for each ERCC spike-in group
}
\description{
Calculate the median expression for each ERCC group
}
