% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{camera_table}
\alias{camera_table}
\title{Convert camera output into a datatable}
\usage{
camera_table(results)
}
\arguments{
\item{results}{data.frame, the output of the \code{\link[limma:camera]{limma::camera()}} function}
}
\value{
a datatable htmlwidget
}
\description{
Convert camera output into a datatable
}
