% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{ercc}
\alias{ercc}
\title{Concentrations for ERCC spike-in mixes}
\usage{
ercc()
}
\value{
data.frame with one row for each ERCC spike-in transcript
}
\description{
Concentrations for ERCC spike-in mixes
}
\examples{
head(ercc())
}
