% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_fibroblast_conc}
\alias{fit_fibroblast_conc}
\title{Modeling the SUPT4H1 in fibroblasts effect separately for each siRNA conc.}
\usage{
fit_fibroblast_conc(dge, normalize = c("spike-in", "global"))
}
\arguments{
\item{dge}{DGEList with raw counts}

\item{normalize}{Scalar string, calculate a TMM scaling factor using either
\code{spike-in} or \code{global} normalization.}
}
\value{
A list of \link[limma:MArrayLM]{limma::MArrayLM} fit objects, one for each siRNA
concentration.
}
\description{
Differential expression analysis using functions and methods from the
\code{edgeR} and \code{limma} packages':
\enumerate{
\item The \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}} method is called to determine which genes have
sufficiently large counts to be retained in a statistal analysis, based on
the experimental design.
\item The \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} method is used to calculate a TMM scaling
factor, either based on all endogenous genes or only on the external ERCC
spike-ins retained after filtering, as determined by the \code{normalize}
argument.
\item The \code{\link[limma:voom]{limma::voom()}} function is used to transform raw count data to
log2-counts per million (logCPM), estimate the mean-variance
relationship and compute appropriate observation-level weights.
\item A linear model is fit for each gene with the \code{\link[limma:lmFit]{limma::lmFit()}} function.
\item Moderated t-statistics and log-odds of differential expression are
calculated with the \code{\link[limma:eBayes]{limma::eBayes()}} function using an empirical Bayes
approach.
A linear mixed model is fit by REML individually for each gene, taking
account the correlation between technical replicates performed within each
cell line.
}
}
