% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_fibroblasts}
\alias{fit_fibroblasts}
\title{Fit a mixed linear model to RNA-seq data from fibroblasts}
\usage{
fit_fibroblasts(dge, normalize = c("spike-in", "global"),
  output = c("fit", "v"))
}
\arguments{
\item{dge}{DGEList with raw counts}

\item{normalize}{Scalar string, calculate a TMM scaling factor using either
\code{spike-in} or \code{global} normalization.}

\item{output}{Scalar character specifying which output to return, either \code{v}
to return output of \code{\link[limma:voom]{limma::voom()}} or \code{fit} to return the output of
\code{\link[limma:eBayes]{limma::eBayes()}}.}
}
\value{
\link[limma:MArrayLM]{limma::MArrayLM} fit object
}
\description{
Differential expression analysis using functions and methods from the
\code{edgeR} and \code{limma} packages.
}
\details{
\enumerate{
\item The \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr()}} method is called to determine which genes have
sufficiently large counts to be retained in a statistal analysis, based on
the experimental design.
\item The \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} method is used to calculate a TMM scaling
factor, either based on all endogenous genes or only on the external ERCC
spike-ins retained after filtering, as determined by the \code{normalize}
argument.
\item The \code{\link[limma:voom]{limma::voom()}} function is used to transform raw count data to
log2-counts per million (logCPM), estimate the mean-variance
relationship and compute appropriate observation-level weights.
\item The \code{\link[limma:duplicateCorrelation]{limma::duplicateCorrelation()}} function is used to estimate the
correlation between between technical replicates (e.g. within each
cell line).
\item The \code{\link[limma:voom]{limma::voom()}} function is called a second time, this time also
providing the average estimated inter-duplicate correlation determined
above.
\item A linear model is fit for each gene with the \code{\link[limma:lmFit]{limma::lmFit()}} function.
\item Moderated t-statistics and log-odds of differential expression are
calculated with the \code{\link[limma:eBayes]{limma::eBayes()}} function using an empirical Bayes
approach.
A linear mixed model is fit by REML individually for each gene, taking
account the correlation between technical replicates performed within each
cell line.
}
}
\examples{
library(limma)
# global shift is detected with spike-in normalization
fit_spikein <- fit_fibroblasts(normalize = "spike-in")
limma::plotMA(fit_spikein, coef = "SUPT4H1",
    main = "Fibroblasts: SUPT4H1 RNAi vs Scramble (spike-in normalization)",
    bg.col = adjustcolor("grey", 0.3), bg.cex = 0.7,
    status = factor(fit_spikein$genes$ercc_logFC),
    value = c("-2", "0", "0.58", "1"))
abline(h = 0, lty = 2)
# but not with global normalization
fit_global <- fit_fibroblasts(normalize = "global")
limma::plotMA(fit_global, coef = "SUPT4H1",
    main = "Fibroblasts: SUPT4H1 RNAi vs Scramble (global normalization)",
    bg.col = adjustcolor("grey", 0.3), bg.cex = 0.7,
    status = factor(fit_spikein$genes$ercc_logFC),
    value = c("-2", "0", "0.58", "1"))
abline(h = 0, lty = 2)
}
