% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{human_numbers}
\alias{human_numbers}
\title{Format numbers in a human-friendly format as axis labels}
\usage{
human_numbers(x, signif = 1)
}
\arguments{
\item{x}{Vector of numbers to be reformatted}

\item{signif}{Scalar integer, the number of significant digits to include}
}
\value{
Character vector
}
\description{
Format numbers in a human-friendly format as axis labels
}
\examples{
human_numbers(c(-100, 1, 1000, 1.5e6))
}
