% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_dataset}
\alias{load_dataset}
\title{Retrieve raw abundance data from NCBI GEO}
\usage{
load_dataset(experiment = c("HEK293", "Fibroblasts", "HEK293_71D"))
}
\arguments{
\item{dataset}{Scalar character, the experiment of interest, either \code{HEK293}
or \code{Fibroblasts}.}
}
\value{
\link[edgeR:DGEList]{edgeR::DGEList} object
}
\description{
Retrieve raw abundance data from NCBI GEO
}
\note{
An internet connection is required to retrieve the processed counts
from NCBI's GEO repository
}
