% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_flow_data}
\alias{load_flow_data}
\title{Load flow cytometry results}
\usage{
load_flow_data(experiment = c("siRNA", "RNAse"))
}
\arguments{
\item{experiment}{Scalar character, one of \code{RNAi} or \code{RNAse}, specifying the
results from cells treated with SUPT4H1 / control siRNAs or RNAse,
respectively.}
}
\value{
data.frame with columns \code{signal} and \code{condition}.
}
\description{
Load flow cytometry results
}
\examples{
head(load_flow_data("siRNA"))
}
