% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_intron_counts}
\alias{load_intron_counts}
\title{Load precomputed table of intronic counts for HEK293 cells}
\usage{
load_intron_counts()
}
\value{
\link[edgeR:DGEList]{edgeR::DGEList} object
}
\description{
Load precomputed table of intronic counts for HEK293 cells
}
