% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{peptide_chain_elongation}
\alias{peptide_chain_elongation}
\title{Ensembl identifiers for Peptide Chain Elongation Reactome gene set}
\usage{
peptide_chain_elongation()
}
\value{
list with a single element, a named vector with ensembl identifiers
annotated in the \code{Peptide Chain Elongation} Reactome gene set.
}
\description{
Entrez gene identifiers from the \code{Peptide Chain Elongation Reactome} gene
set were curated from the \code{reactome.db} Bioconductor package
(version 1.64.0) and mapped to ensembl gene identifiers using the
\code{org.Hs.eg.db} Bioconductor package (version 3.6.0).
}
\examples{
peptide_chain_elongation()
}
