% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_biotypes}
\alias{plot_biotypes}
\title{Plot SUPT4H1 RNAi effect by biotype}
\usage{
plot_biotypes(tt)
}
\arguments{
\item{tt, }{data.frame, typically the output of the \code{\link[limma:topTable]{limma::topTable()}}
function}
}
\value{
ggplot object
}
\description{
Plot SUPT4H1 RNAi effect by biotype
}
