% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_boxes}
\alias{plot_boxes}
\title{Generate a boxplot from a topTable}
\usage{
plot_boxes(tt, colors = brewer.pal(n = 8, "Set1"), ylim = c(-3, 2),
  labels = FALSE, nudge_x = 0.2, nudge_y = 0.25, nudge_spt4 = 0.25,
  cex = 4, show_spt4 = FALSE)
}
\arguments{
\item{tt, }{data.frame, typically the output of the \code{\link[limma:topTable]{limma::topTable()}}
function}

\item{colors}{Character vector with colors for each box}

\item{ylim}{Numeric vector of length two that specifies the y-axis limits}

\item{nudge_x}{Scalar numeric, the x-axis offset for the labels}

\item{nudge_y}{Scalar numeric, the y-axis offset for the labels}

\item{nudge_spt4}{Scalar numeric, the y-axis offset of the SPT4H1 label}

\item{cex}{Scalar numeric, character expansion for the SUPT4H1 shape}

\item{show_spt4}{Scalar logical, highlight the log-fold change for SUPT4H1?}
}
\value{
A ggplot object
}
\description{
Generate a boxplot from a topTable
}
