% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot the normalized CPMs for a single gene symbol}
\usage{
plot_expression(gene_symbol, dge, labels)
}
\arguments{
\item{gene_symbol}{An HGNC gene symbol, must match an entry in
\code{dge$genes$hgnc_symbol}.}

\item{dge}{A DGEList list object with appropriate normalization applied}

\item{labels}{A named character vector that will be used to rename elements
of the \code{dge$samples$treatment} and \code{dge$samples$cell_line} columns.}
}
\value{
A ggplot object
}
\description{
Plot the normalized CPMs for a single gene symbol
}
