% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_flow_density}
\alias{plot_flow_density}
\title{Create density plots of flow cytometry results}
\usage{
plot_flow_density(x, red = "SUPT4H1", grey = "Scramble")
}
\arguments{
\item{x}{data.frame with two columns:
\itemize{
\item signal
\item condition
}}

\item{red}{Scalar character, the condition displayed in red.}

\item{grey}{Scalar character, the condition displayed in grey.}
}
\description{
Create density plots of flow cytometry results
}
\examples{
plot_flow_density(load_flow_data("siRNA"))
}
