% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot differential expression statistics for a gene set}
\usage{
plot_gsea(tt, gene_sets, id, xlim = c(-3, 3), y = c("logFC", "y"))
}
\arguments{
\item{tt, }{data.frame, typically the output of the \code{\link[limma:topTable]{limma::topTable()}}}

\item{gene_sets}{a list of one or more gene sets, each with ensembl
identifiers}

\item{xlim}{Numeric vector of length two, the x-axis limits}

\item{y}{Scalar character specifying which statistic to plot.}
}
\value{
ggplot object
}
\description{
Plot differential expression statistics for a gene set
}
