% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ma}
\alias{plot_ma}
\title{Generate an MA plot from a topTable}
\usage{
plot_ma(tt, xlim = c(-3, 12), ylim = c(-6, 4), nudge_spt4 = 0.5)
}
\arguments{
\item{tt, }{data.frame, typically the output of the \code{\link[limma:topTable]{limma::topTable()}}
function}

\item{xlim}{Numeric vector of length two that specifies the x-axis limits}

\item{ylim}{Numeric vector of length two that specifies the y-axis limits}

\item{nudge_spt4}{Scalar numeric, the y-axis offset of the SPT4H1 label}
}
\value{
A ggplot object
}
\description{
Generate an MA plot from a topTable
}
