% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot normalized CPMs for a single human gene}
\usage{
plot_scatter(dge, gene_symbol = "GFP", x = "rnai_treatment")
}
\arguments{
\item{dge}{A DGEList list object with appropriate normalization applied}

\item{gene_symbol}{Scalar character, a human gene identifier}
}
\value{
ggplot object
}
\description{
Plot normalized CPMs for a single human gene
}
