% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{reactome_sets}
\alias{reactome_sets}
\title{Return a list of ensembl identifiers for each Reactome pathway}
\usage{
reactome_sets(min_size = 10, max_size = Inf)
}
\arguments{
\item{min_size}{Scalar integer, the minimum size of each set}

\item{max_size}{Scalar integer, the maximum size of each set}
}
\description{
Return a list of ensembl identifiers for each Reactome pathway
}
\examples{
gene_sets <- reactome_sets()
gene_sets[1:2]
}
